Refactor Styled Components and add testing
