+++ src/styledComponents/Star/index.js
