+++ src/styledComponents/StyledGif/index.js
