+++ src/styledComponents/index.js
@@ -1,9 +1,13 @@
 import Message from './Message';
 import Page from './Page';
 import Image from './Image';
+import Star from './Star';
+import StyledGif from './StyledGif';
 
 export {
   Image,
   Page,
   Message,
+  Star,
+  StyledGif,
 };
